Removes SimpsonCharacterService interface
