+++ src/main/java/com/javachallengers/simpson/service/SimpsonCharacterService.java
@@ -1,32 +1,47 @@
 package com.javachallengers.simpson.service;
 
+import java.util.Objects;
 import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import com.javachallengers.simpson.exception.SimpsonCharacterException;
 import com.javachallengers.simpson.model.SimpsonCharacter;
 import com.javachallengers.simpson.model.dto.SimpsonCharacterRequestDTO;
+import com.javachallengers.simpson.repository.SimpsonCharacterRepository;
 
-/**
- * Service with business rules related to Simpson Characters
- */
-public interface SimpsonCharacterService {
-	/**
-	 * Gets an Simpson Character by its unique id.
-	 * 
-	 * @param id character id
-	 * 
-	 * @return the character with the given id. {@code Optional#empty()} if not found.
-	 */
-	Optional<SimpsonCharacter> getCharacterById(String id);
+@Service
+public class SimpsonCharacterService {
+	private SimpsonCharacterRepository simpsonCharacterRepository;
+	
+	@Autowired
+	public SimpsonCharacterService(SimpsonCharacterRepository simpsonCharacterRepository) {
+		super();
+		this.simpsonCharacterRepository = simpsonCharacterRepository;
+	}
+	
+	public Optional<SimpsonCharacter> getCharacterById(String id) {
+		Objects.requireNonNull(id, "id must not be null");
+		return simpsonCharacterRepository.findById(id);
+	}
 
-	/**
-	 * Creates a new Simpson Character
-	 * 
-	 * @param simpsonCharacterRequestDTO
-	 * 
-	 * @return Created {@link SimpsonCharacter} entity
-	 * 
-	 * @throws SimpsonCharacterException in case of validation error (e.g. invalid name/surname) or creation failure (e.g. database error).
-	 */
-	SimpsonCharacter createNewCharacter(SimpsonCharacterRequestDTO simpsonCharacterRequestDTO) throws SimpsonCharacterException;
+	public SimpsonCharacter createNewCharacter(SimpsonCharacterRequestDTO simpsonCharacterRequestDTO) throws SimpsonCharacterException {
+		Objects.requireNonNull(simpsonCharacterRequestDTO, "");
+		
+		String name = simpsonCharacterRequestDTO.getName();
+		if(StringUtils.isEmpty(name)) {
+			throw new SimpsonCharacterException("name must be provided");
+		}
+		
+		String surname = simpsonCharacterRequestDTO.getSurname();
+		if(StringUtils.isEmpty(surname)) {
+			throw new SimpsonCharacterException("surname must be provided");
+		}
+		
+		SimpsonCharacter simpsonCharacter = new SimpsonCharacter(name, surname);
+		
+		return simpsonCharacterRepository.save(simpsonCharacter);
+	}
 }
